+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
+* Method `setSettings` of `UploadClientInterface` was renamed to `updateSettings`.
 * Methods (`request`, `base`, `fromUrl`, `fromUrlStatus`, `group`, 
 `groupInfo`, `info`, `multipartStart`, `multipartUpload`, 
 `multipartComplete`, `multipart`, `fileFrom`, `groupFrom`) was exported 
@@ -18,6 +19,8 @@ from `index.ts` to make better tree shaking.
 
 ### Removed
 
+* `addUpdateSettingsListener` and `removeUpdateSettingsListener` from 
+`UploadClientInterface`.
 * Removed `from` param from `fileFrom` and `groupFrom`. 
 * Removed `FileFromEnum` and `GroupFromEnum`.